Relocate integration-ey bits of HeliosSoloDeploymentTest -> HeliosSoloDeploymentIT
